requirements.txt
tensorflow==2.10.1
keras==2.10.0

numpy==1.23.5
Pillow==10.0.0

gradio==3.44.0
gradio-client==1.0.0

huggingface-hub==0.17.3
datasets==2.14.0

pydantic==2.10.6
click==8.0.4
protobuf<4

torch
transformers
matplotlib
reportlab
